server: Add VP8 support to the GStreamer encoder
The Spice server administrator can specify the preferred encoder and
codec preferences to optimize for CPU or bandwidth usage. Preferences
are described in a semi-colon separated list of encoder:codec pairs.
For instance 'gstreamer:vp8;spice:mjpeg' to pick first the GStreamer
VP8 video encoder and the original MJPEG video encoder as a fallback.
The server has a default preference list which can also be selected by
specifying 'auto' as the preferences list.

The server then picks a codec supported by the client based on the
following new client capabilities:
 * SPICE_DISPLAY_CAP_MULTI_CODEC which denotes a recent client that
   supports multiple codecs. This capability is needed to not have to
   hardcode that MJPEG is supported. This makes it possible to write
   clients that don't support MJPEG.
 * SPICE_DISPLAY_CAP_CODEC_XXX, where XXX is a supported codec, for now
   MJPEG and VP8.

Signed-off-by: Francois Gouget <fgouget@codeweavers.com>