server: Avoid copying the input frame in the GStreamer encoder
This relies on the GStreamer buffer's lifetime being short enough which
it is because:
 - We encode frames one by one.
 - For all encoders but MJPEG, the first element of the pipeline will
   convert the bitmap to another image format which entails copying it.
   So by the time the encoder starts its work, this buffer will not be
   needed anymore.
 - The MJPEG encoder does not perform inter-frame compression and thus
   does not need to keep hold of this buffer once it has processed it.

Note that we can only avoid copies for the first 1 Mpixels or so.
That's because Spice splits larger frames into more chunks than we can
fit GstMemory fragments in a GStreamer buffer. So if there are more
pixels we will avoid copies for the first 3840 KB and copy the rest.

Signed-off-by: Francois Gouget <fgouget@codeweavers.com>