server: Remove the width and height parameters of encode_frame()
They always match the size of the source bitmap area, even for sized
streams and thus are redundant.
Note also that we do not lose any flexibility since the source area
parameter can be used to only encode part of the bitmap (although
that's not a feature we currently use).

Signed-off-by: Francois Gouget <fgouget@codeweavers.com>